requirements.txt
torch==2.0.0
transformers==4.25.1
langchain==0.1.13
openai==1.11.0
whisper-timestamped==1.12.20
ffmpeg-python==0.2.0
langchain-community==0.0.29
langchain-core==0.1.33
langchain_openai==0.0.8
langchain-experimental==0.0.52
spacy==3.7.2
fastapi==0.75.0
yt_dlp==2024.3.10
python-dotenv==1.0.0
faiss-cpu==1.8.0